// src/lib/schema.ts
export const PatientFormSchema = z.object({
    first_name: z
        .string()
        .trim()
        .min(2, "First name must be at least 2 characters")
        .max(30, "First name can't be more than 50 characters"),
    last_name: z
        .string()
        .trim()
        .min(2, "dLast name must be at least 2 characters")
        .max(30, "First name can't be more than 50 characters"),
    date_of_birth: z.coerce.date(),
    gender: z.enum(["MALE", "FEMALE"], { message: "Gender is required" }),

    phone: z.string().min(10, "Enter phone number").max(10, "Enter phone number"),
    email: z.string().email("Invalid email address."),
    address: z
        .string()
        .min(5, "Address must be at least 5 characters")
        .max(500, "Address must be at most 500 characters"),
    marital_status: z.enum(["married", "single", "divorced", "widowed", "separated"], {
        message: "Marital status is required."
    }),
    emergency_contact_name: z
        .string()
        .min(2, "Emergency contact name is required.")
        .max(50, "Emergency contact must be at most 50 characters"),
    emergency_contact_number: z
        .string()
        .min(10, "Enter phone number")
        .max(10, "Enter phone number"),
    relation: z.enum(["mother", "father", "husband", "wife", "other"], {
        message: "Relations with contact person required"
    }),
    blood_group: z.string().optional(),
    allergies: z.string().optional(),
    medical_conditions: z.string().optional(),
    medical_history: z.string().optional(),
    insurance_provider: z.string().optional(),
    insurance_number: z.string().optional(),
    privacy_consent: z
        .boolean()
        .default(false)
        .refine((val) => val === true, {
            message: "You must agree to the privacy policy."
        }),
    service_consent: z
        .boolean()
        .default(false)
        .refine((val) => val === true, {
            message: "You must agree to the terms of service."
        }),
    medical_consent: z
        .boolean()
        .default(false)
        .refine((val) => val === true, {
            message: "You must agree to the medical treatment terms."
        }),
    img: z.string().optional()
})